requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.8
django-cors-headers==3.0.1
django-debug-toolbar==1.11
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
drf-extensions==0.5.0
idna==2.8
itsdangerous==1.1.0
kombu==4.5.0
Pillow==6.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
qiniu==7.2.4
qrcode==6.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0


